requirements.lock.txt
aiosqlite==0.22.1
alembic==1.18.4
altair==6.0.0
amplitude-analytics==1.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
apprise==1.9.9
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==26.1.0
beartype==0.22.9
beautifulsoup4==4.14.3
black==25.12.0
bleach==6.3.0
blinker==1.9.0
brotli==1.2.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coolname==4.1.0
coverage==7.13.5
cronsim==2.7
cryptography==46.0.6
cyclopts==4.10.1
dateparser==1.4.0
distlib==0.4.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
duckdb==1.5.1
-e git+https://github.com/wobe8333-hub/E_Commerce.git@978c6c13c675720895eca0b22f6c3640384cefeb#egg=e_commerce
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fakeredis==2.34.1
fastapi==0.135.2
feedparser==6.0.10
filelock==3.25.2
fsspec==2026.3.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.28.1
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.3.0
griffe==2.0.2
griffecli==2.0.2
griffelib==2.0.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
identify==2.6.18
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaggle==1.8.3
kagglesdk==0.1.14
librt==0.7.7
lightgbm==4.6.0
loguru==0.7.3
lupa==2.6
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mutagen==1.47.0
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.18.1
nodeenv==1.10.0
numpy==2.4.0
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.40.0
orjson==3.11.7
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pendulum==3.2.0
Pillow==10.1.0
platformdirs==4.5.1
playwright==1.57.0
plotly==6.6.0
pluggy==1.6.0
polars==1.39.3
polars-runtime-32==1.39.3
pre_commit==4.5.1
prefect==3.6.24
prometheus_client==0.24.1
proto-plus==1.27.0
protobuf==6.33.2
py-key-value-aio==0.4.4
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybit==5.7.0
pycparser==3.0
pycryptodome==3.23.0
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydocket==0.18.2
pyee==13.0.0
Pygments==2.20.0
pyparsing==3.3.1
pytest==8.2.0
pytest-cov==7.1.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-dotenv==1.0.0
python-json-logger==4.1.0
python-multipart==0.0.6
python-slugify==8.0.4
pytokens==0.3.0
pytrends==4.9.2
pytz==2025.2
pywin32==311
PyYAML==6.0.3
readchar==4.2.1
redis==7.4.0
referencing==0.37.0
regex==2026.3.32
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
ruff==0.15.8
scipy==1.17.1
semver==3.0.4
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.48
starlette==1.0.0
streamlit==1.55.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.5
tqdm==4.67.1
typer==0.24.1
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uncalled-for==0.2.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.24.0
virtualenv==21.2.0
watchdog==6.0.0
watchfiles==1.1.1
webencodings==0.5.1
websocket-client==1.7.0
websockets==15.0.1
win32_setctime==1.2.0
yt-dlp==2024.3.10
zipp==3.23.0
